Sanitized and validated

+++ control_card_delete.php
@@ -2,7 +2,14 @@
 
 $HSDB = require './lib/connect.php';
 
-$id = filter_input(INPUT_POST, 'id');
+$idOptions = array('options'=>array('default'=>0, 'min_range'=>1));
+
+$id = filter_input(INPUT_POST, 'id', FILTER_VALIDATE_INT, $idOptions);
+if ($id == 0) {
+    echo "PLEASE USE A CORRECT ID.";
+    exit();
+}
+
 $card = $HSDB->getCardById($id)[0];
 $HSDB -> deleteCard($id);
